==> analysis/integration/merfish_mc/02.merge_and_pca.ipynb <==
from ALLCools.clustering import *
from sklearn.decomposition import PCA
import anndata

import numpy as np

n_train_cell = 100000

## Merge Adata

ref_adata = anndata.read_h5ad('./adata/mc_input.h5ad')
query_adata = anndata.read_h5ad('./adata/merfish_input.h5ad')

use_var = ref_adata.var_names.intersection(query_adata.var_names)
ref_adata._inplace_subset_var(use_var)
query_adata._inplace_subset_var(use_var)

adata_merge = ref_adata.concatenate(query_adata,
                                    batch_categories=['mC', 'merfish'],
                                    batch_key='Modality',
                                    index_unique=None)
adata_merge

np.random.seed(0)

# select mC cells to fit the model
train_cell = np.zeros(ref_adata.shape[0]).astype(bool)
if ref_adata.shape[0] > n_train_cell:
    train_cell[np.random.choice(np.arange(ref_adata.shape[0]), n_train_cell,
                                False)] = True
else:
    train_cell[:] = True

ref_adata.obs['Train'] = train_cell.copy()

## Run PCA on Merged Adata

if ref_adata.X[ref_adata.obs['Train'].values].shape[1] >= 100:
    model = PCA(n_components=100, svd_solver='arpack', random_state=0)
else:
    model = PCA(n_components=ref_adata.X[ref_adata.obs['Train'].values].shape[1]-1, svd_solver='arpack', random_state=0)

# use selected mC cells to fit
model.fit(ref_adata.X[ref_adata.obs['Train'].values])
sel_dim = (model.singular_values_ != 0)
print(sel_dim.sum())

# transform all other data
chunk_size = 50000
chunks = []
for chunk_start in range(0, adata_merge.shape[0], chunk_size):
    chunks.append(
        model.transform(adata_merge.X[chunk_start:(chunk_start + chunk_size)]))

adata_merge.obsm['X_pca'] = np.concatenate(chunks, axis=0)[:, sel_dim]
adata_merge.obsm['X_pca'] /= model.singular_values_[sel_dim]

significant_pc_test(adata_merge, p_cutoff=0.1, obsm='X_pca')

for m in ['mC', 'merfish']:
    adata = adata_merge[adata_merge.obs['Modality'] == m].copy()
    adata.write_h5ad(f'./adata/{m.lower()}_pca.h5ad')

adata_merge
